geoimagenet_api/geoserver_setup/requirements.txt
click
loguru
pyyaml
dataclasses
git+https://github.com/boundlessgeo/gsconfig#egg=gsconfig
